faces.py: PropFace formats better dist, support.
For dist and support, we get it from node.dist and node.support now.

This makes that any value read as a string (like support if it is
specified in the NHX entended attributes in a newick), will be still
seen as a float when accessing it (and formatting it with "%.2g" and
such).