Add pixi backend to draw sequences.
This is just a draft.

It works at this moment by drawing sequences on a pixi canvas on the
aligned panel, and everything else adding svgs as before.

But we want to have the ability to draw with pixi anywhere, and also
in the circular representation. This will need some deeper changes.