Fix sequences for svg and pixi, in rectangular and circular mode.
We were not going well from imin to imax (their values were not
correct and we were drawing elements that were not shown).

Now we only draw the elements in the sequence that are visible (in
rectangular mode). In circular mode, we draw always all the elements
(it is tricky to remove the ones that would be outside of the limits).

But the biggest change is that now pixi sequences work well for
circular mode too.