Draw aligned items in different panels.
This change makes that the column in
  Decoration(..., position='aligned', column=...)
will correspond to a different panel (starting from 1, since panel 0 is
the main tree panel).

This means that all the drawings from all the layouts that go to
different panels will be globally aligned.

This works in rectangular and circular mode. And it prepares the way
to have headers and footers there too.