Remove unused .travis.yml
We do not use travis, and have not used for a long time.

Travis stopped supporting foss projects a few years ago:
https://earthly.dev/blog/migrating-from-travis/

We use instead github actions now (which may be a problem in
the future if they remove their "GHA open source plan").
See the relevant files in the .github directory.