Remove Decorations.
After our group meeting, we decided to remove the Decorations, and
have the information about position, column and anchor inside the
Faces themselves. Yay!

The Face base class now is a bit less silly, and simply sets those
location values.

This commit includes all the relevant changes. And with the commit that
I am about to make in ete-gallery, the examples there will be updated too.